Soccer Player Tracking Final/requirements.txt
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.19.0
Pillow>=9.0.0
PyYAML>=6.0
opencv-python>=4.5.0
filterpy>=1.4.5
scipy>=1.7.0 